requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
langchain>=0.1.0
python-dotenv>=1.0.0
pypdf>=3.9.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
transformers>=4.30.0
torch>=2.0.0
together>=0.2.5
langgraph>=0.0.15
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-jose>=3.3.0
pandas>=2.0.0
nltk>=3.8.1
networkx>=3.0
matplotlib>=3.7.0